pyproject.toml
[tool.poetry]
name = "ai-travel-planner-agent"
version = "0.1.0"
description = ""
authors = ["Sai Suchir Tyada <styada3@gatech.edu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
langgraph = "*"
langchain = "*"
langchain-core = "*"
langchain-ollama = "*"
langchain-google-genai = "*"
langchain-community = "*"
duckduckgo-search = "*"
ddgs = "*"
pydantic = "*"
python-dotenv = "*"
fastapi = "^0.129.0"
uvicorn = "^0.41.0"
certifi = "^2026.1.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
